+++ src/test/java/coupon/application/coupon/CouponServiceTest.java
@@ -1,4 +1,4 @@
-package coupon.application;
+package coupon.application.coupon;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
